requirements.txt
ansible==2.10.5
ansible-base==2.10.5
appnope==0.1.2
asgiref==3.3.1
asn1crypto==1.4.0
astroid==2.4.2
autopep8==1.5.4
backcall==0.2.0
bcrypt==3.2.0
cffi==1.14.4
click==7.1.2
cryptography==3.3.1
decorator==4.4.2
Django==3.1.5
django-appconf==1.0.4
django-autoslug==1.9.8
django-imagekit==4.0.2
Faker==5.6.5
Flask==1.1.2
gunicorn==20.0.4
ipython==7.13.0
ipython-genutils==0.2.0
isort==5.7.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
lazy-object-proxy==1.4.0
MarkupSafe==1.1.1
mccabe==0.6.1
mpmath==1.1.0
numpy==1.19.5
packaging==20.8
paramiko==2.7.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
pilkit==2.0
Pillow==8.1.0
prompt-toolkit==3.0.13
psycopg2==2.8.6
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.7.4
pylint==2.6.0
PyNaCl==1.4.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
PyYAML==5.4.1
scipy==1.6.0
simplegeneric==0.8.1
six==1.15.0
sqlparse==0.4.1
sympy==1.7.1
text-unidecode==1.3
toml==0.10.2
traitlets==5.0.5
typed-ast==1.4.2
Unidecode==1.1.2
uWSGI==2.0.19.1
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
yolk3k==0.9
